Create test
Your task is to complete the program. Write a Client class that has two constructors and two methods: returnName and returnID. Client class has three String type attributes: name, ID and address. In addition you should write the classes CashClient and AccountClient which derive from Client class. CashClient class has one method: payCash which receives a sum as parameter and prints the text shown in the example print. AccountClient has one attribute that keeps track of the balance on the client account. In addition the AccountClient class has two methods: deposit which adds the given number to the account balance and withdraw which subtracts the given number from the account balance. AccountClient class also requires toString method which prints the name, ID and account balance of a client.

Program to complete(no need to copy this to the return field):

public class ClientRegister {
    public static void main(String args[]) {
        CashClient client1 = new CashClient("Lisa", "101070-1111");
        AccountClient client2 = new AccountClient("Benny", "101070-1111");
        AccountClient client3 = new AccountClient("Barney", "040474-1571", "Someroad 3");
        Client client4 = new Client("Client", "111170-1111", "Road 4"); 
           
        client1.payCash(70);
        client2.deposit(100);
        client3.deposit(500);
        client2.withdraw(20);
        
        System.out.println(client2);
        System.out.println(client3);
    }
}

Note: In this exercise the solution is part of a conversion unit where many classes have been declared. Because of this the classes are not declared as public using the public attribute. 


Example output:
Lisa(101070-1111): Amount paid 70 euros.
Benny(101070-1111): 80.0
Barney(040474-1571): 500.0
The verification of program output does not account for whitespace and is not case-sensitive (the least strict comparison level)