Create Exceptions
Your task is to complete the program by writing two classes (SmallException and BigException) and two methods to the ownException class (printErrorReport and testValue). SmallException and BigException are individual exception classes deriving from Exception class. Both classes require a constructor that receives a String object as parameter. Parameter is used to relay an informative message with the exception. Parameter is relayed to the superclass constructor. printErrorReport method receives the exception as a parameter and prints the error report of the exception using getMessage method. testValue method receives the tested number as a parameter. If the number is lower than five, method throws the SmallException and parameter is the message: Value is lower than 5. If the number is higher than 10, method throws the BigException and parameter is the message: Value is higher than 10.

Program to complete(no need to copy this to the return field):

import java.util.Scanner;
import java.util.*;

public class ownException {
    public static void main(String[] args) {
        Scanner reader = new Scanner(System.in);
        int testedNumber = 0;
        boolean inputCorrect = true;
        do {
            try {
                System.out.print("Type in the number for testing: ");
                testedNumber = reader.nextInt();
                inputCorrect = true;
            } catch (InputMismatchException e) {
                System.out.println("You did not type in an integer!");
                inputCorrect = false;
                reader.nextLine();
            }
        } while(!inputCorrect);
        
        try {
            testValue(testedNumber);
            System.out.println("Value was between 5 and 10.");
        } catch(SmallException e) {
            System.out.println("SmallException caught!");
            printErrorReport(e);
        } catch(BigException e) {
            System.out.println("BigException caught!");
            printErrorReport(e);
        }          
    }

    // Your code here

Note: Remember to write the wave bracket to end the ownException class.  In this exercise the solution is part of a conversion unit where many classes have been declared. Because of this the classes are not declared as public using the public attribute. 


Example output:
Type in the number for testing: a
You did not type in an integer!
Type in the number for testing: 6
Value is between 5 and 10.
The verification of program output does not account for whitespace and is not case-sensitive (the least strict comparison level)